+++ pfbex/metadata.py
@@ -4,4 +4,4 @@
 APP_NAME = 'pfbex'
 
 # Application version
-APP_VERSION = '0.3.0'
+APP_VERSION = 'develop'